# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Aggregato API
  version: 1.0.0
  description: |
    Single-user, self-hosted media log aggregator. Every endpoint requires authentication by
    default: either `Authorization: Bearer <api.token>` or the session cookie issued by
    `POST /auth/session`. When `api.allow_unauthenticated_readonly` is enabled, credential-free
    GET/HEAD requests are allowed as read-only requests; writes and websockets still require a
    credential. Backups, retained diagnostics, and ingest-failure payloads require operator
    access even though they use GET. Public and read-only credentials receive 403 on those routes.

    An optional second credential, `api.readonly_token`, authenticates reads only: any
    POST/PUT/PATCH/DELETE it presents — directly or through the session cookie it exchanges
    itself for — is refused with 403 `readonly-token`. `GET /auth/session` reports which
    credential the caller holds.

    Conventions:
      * Timestamps are ISO 8601 with offset. `logged_at` is always accompanied by
        `logged_precision`; clients must not render a fabricated time as exact.
      * Ratings return raw value, scale, and normalized value. Normalized values are
        comparable within a scale, not across platforms.
      * Pagination is keyset-only via opaque `cursor`. There is no offset parameter.
      * Collection pages use `limit` 1–200, defaulting to 50; out-of-range values are clamped.
...
        is no arbitrary matching-result cap.
...
servers:
...
security:
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /providers/{id}/runs:
    get:
      tags: [providers]
      parameters:
        - $ref: "#/components/parameters/providerId"
        - { name: status, in: query, schema: { $ref: "#/components/schemas/RunStatus" } }
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: Sync run history, newest first, grouped by retry lineage. Provider check runs are excluded; use `Provider.last_check` for the latest diagnostic. Pages are keyset-paginated and capped by the shared limit parameter.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - properties:
                      items: { type: array, items: { $ref: "#/components/schemas/SyncRun" } }
  /providers/{id}/runs/{run_id}/diagnostics:
    get:
      tags: [operations]
      summary: Read retained diagnostics for one sync attempt
      description: Operator access required. Public and read-only credentials are refused.
      parameters:
        - $ref: "#/components/parameters/providerId"
        - { name: run_id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: Retained child log and HTTP response metadata without bodies or headers
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncRunDiagnostics" }
        "403": { $ref: "#/components/responses/Problem" }
  /sync/status:
    get:
      tags: [operations]
      summary: Read the current sync queue and latest run state
      description: >-
        Returns scheduler state plus the newest non-diagnostic run for each provider. While a run is
        open, its phase, counts, checkpoints, and cursor checkpoint are updated in place. This is
        also the initial/reconnect shape sent by the authenticated `/api/v1/ws/sync` websocket.
      parameters:
        - name: provider_id
          in: query
          schema: { type: string }
        - name: lineage_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Current durable sync state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncSnapshot" }
...
  /ingest-failures:
    get:
      tags: [operations]
      summary: Read retained conversion failures
      description: Operator access required because payloads can contain private platform data.
      parameters:
        - $ref: "#/components/parameters/provider"
        - { name: resolved, in: query, schema: { type: boolean, default: false } }
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: Records that failed conversion, with retained provider payloads for diagnosis and replay. Payloads may contain private platform data and are subject to local retention.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - properties:
                      items: { type: array, items: { $ref: "#/components/schemas/IngestFailure" } }
        "403": { $ref: "#/components/responses/Problem" }
  /ingest-failures/{id}/replay:
    post:
      tags: [operations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "202":
          description: A replay job was queued for the worker; normalization and writing remain behind the child/parent process boundary.
          content:
            application/json:
              schema:
                type: object
                required: [replayed, queued]
                properties:
                  replayed: { type: boolean }
                  queued: { type: boolean }
                  job_id: { type: integer, format: int64, nullable: true }
        "409": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
    sessionCookie: { type: apiKey, in: cookie, name: aggregato_session }
  parameters:
...
    providerId: { name: id, in: path, required: true, schema: { type: string } }
    limit:
      name: limit
      in: query
      description: Requested page size. The server clamps it to 1–200 and defaults to 50.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Opaque keyset cursor from the previous page. There is no offset parameter .
    provider:
      name: provider
      in: query
      schema: { type: array, items: { type: string } }
      explode: true
...
  responses:
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
    Page:
      type: object
      required: [items, next_cursor]
      properties:
        items: { type: array, items: {} }
        next_cursor: { type: string, nullable: true }
...
    ErrorClass:
      type: string
      enum: [auth, rate_limit, transport, parse, structure_changed, blocked, internal]
    RunStatus: { type: string, enum: [running, success, partial, failed] }
    ProviderStatus: { type: string, enum: [disabled, idle, syncing, degraded, misconfigured] }
...
    SyncRun:
      type: object
      required: [id, provider_id, lineage_id, attempt, status, phase, started_at, updated_at,
                 items_seen, items_written, items_failed, checkpoint_count]
      properties:
        id: { type: integer, format: int64 }
        provider_id: { type: string }
        lineage_id: { type: string, format: uuid }
        attempt: { type: integer }
        mode: { type: string, enum: [incremental, full, import, replay, check] }
        status: { $ref: "#/components/schemas/RunStatus" }
        phase:
          type: string
          enum: [starting, checking, replaying, fetching, ingesting, finalizing, finished, failed]
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time }
        items_seen: { type: integer }
        items_written: { type: integer }
        items_failed: { type: integer }
        progress_total: { type: integer, nullable: true, description: Null when the provider cannot announce a total }
        progress_percent: { type: integer, nullable: true, minimum: 0, maximum: 100 }
        checkpoint_count: { type: integer }
        last_checkpoint_at: { type: string, format: date-time, nullable: true }
        cursor_before: { type: object, nullable: true, additionalProperties: true, description: Null for public and read-only clients }
        cursor_after: { type: object, nullable: true, additionalProperties: true, description: Null for public and read-only clients }
        error_class: { $ref: "#/components/schemas/ErrorClass" }
        error_message: { type: string, nullable: true, description: Null for public and read-only clients }
        next_retry_at: { type: string, format: date-time, nullable: true }
        log_excerpt: { type: string, nullable: true, description: Application log excerpt; null for public and read-only clients }
    SyncRunDiagnostics:
      type: object
      required: [run_id, raw_responses]
      properties:
        run_id: { type: integer, format: int64 }
        log: { type: string, nullable: true }
        raw_responses:
          type: array
          items: { $ref: "#/components/schemas/RunHttpResponseMetadata" }
    RunHttpResponseMetadata:
      type: object
      additionalProperties: false
      required: [method, url, status]
      properties:
        method: { type: string, maxLength: 32 }
        url: { type: string, maxLength: 2048, description: Endpoint URL with userinfo, query, and fragment removed }
        status: { type: integer }
    SyncProviderState:
      type: object
      required: [id, enabled, status]
      properties:
        id: { type: string }
        enabled: { type: boolean }
        status: { $ref: "#/components/schemas/ProviderStatus" }
        requested_mode: { type: string, enum: [incremental, full, import, replay, check], nullable: true }
        requested_lineage_id: { type: string, format: uuid, nullable: true }
        next_run_at: { type: string, format: date-time, nullable: true }
    SyncSnapshot:
      type: object
      required: [type, generated_at, providers, runs]
      properties:
        type: { type: string, enum: [snapshot] }
        generated_at: { type: string, format: date-time }
        providers:
          type: array
          items: { $ref: "#/components/schemas/SyncProviderState" }
        runs:
          type: array
          items: { $ref: "#/components/schemas/SyncRun" }
...
    IngestFailure:
      type: object
      properties:
        id: { type: integer, format: int64 }
        provider_id: { type: string }
        sync_run_id: { type: integer, format: int64 }
        native_id: { type: string, nullable: true, description: Immutable provider identity captured for safe replay }
        stage: { type: string, enum: [fetch, validate, normalize, write] }
        error: { type: string }
        raw_payload:
          type: object
          description: Retained provider data for diagnosis/replay; may include private notes, identifiers, or source URLs. Authenticated operators should treat it as sensitive.
        created_at: { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
...